move AcceptResult to its own file
